plug-ins: Fix parameter type in psd_read_len()
The function expects *data to be 64bit wide which is incidentally true
on some architectures but not others. Make it explicit (and portable).

+++ plug-ins/file-psd/psd-util.c
@@ -204,7 +204,7 @@ psd_read (GInputStream  *input,
 
 gboolean
 psd_read_len (GInputStream  *input,
-              gsize         *data,
+              guint64       *data,
               gint           psd_version,
               GError       **error)
 {
